backend/requirements.txt
fastapi
uvicorn
onnxruntime
pillow
numpy

